Seperate Cache from database package for future use
