+++ Apromore-Core-Components/Apromore-Manager/src/main/java/org/apromore/cache/ehcache/CacheRepositoryImpl.java
@@ -22,10 +22,9 @@
  * #L%
  */
 
-package org.apromore.dao.jpa;
+package org.apromore.cache.ehcache;
 
 import org.apromore.exception.CacheException;
-import org.apromore.dao.CacheRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
