+++ Apromore-Core-Components/Apromore-Manager/src/main/resources/META-INF/spring/managerContext.xml
@@ -51,6 +51,7 @@
     <!-- Import other Application Context used to construct the Spring context. -->
     <import resource="manager-osgi-context.xml"/>
     <import resource="managerContext-jpa.xml"/>
+    <import resource="managerContext-cache.xml"/>
     <import resource="managerContext-services.xml"/>
     <import resource="managerContext-scheduler.xml"/>
