+++ Apromore-Core-Components/Apromore-Manager/src/test/java/org/apromore/dao/jpa/CacheRepositoryImplTest.java
@@ -25,7 +25,8 @@ package org.apromore.dao.jpa;
 import junit.framework.Assert;
 import net.sf.ehcache.CacheManager;
 import net.sf.ehcache.Ehcache;
-import org.apromore.dao.CacheRepository;
+import org.apromore.cache.ehcache.CacheRepository;
+import org.apromore.cache.ehcache.CacheRepositoryImpl;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;
